+++ mpltex/acs.py
@@ -14,6 +14,7 @@ in actual physical size rahter than pixel unit.
 from .general import MPLdecorator
 from .colors import default_color_cycler
 from .layout import GOLDEN_RATIO
+from .styles import latex_preamble
 
 __all__ = ['acs_decorator', ]
 
@@ -33,12 +34,8 @@ _params = {'font.family': 'sans-serif',
            'font.sans-serif': ['Helvetica', 'Computer Modern Sans serif'],
            'font.size': 8,
            'text.usetex': True,
-           # To force LaTeX use Helvetica fonts.
-           'text.latex.preamble': [r'\usepackage{siunitx}',
-                                   r'\sisetup{detect-all}',
-                                   r'\usepackage{helvet}',
-                                   r'\usepackage[eulergreek,EULERGREEK]{sansmath}',
-                                   r'\sansmath'],
+           'text.latex.preamble': latex_preamble,  # To force LaTeX use Helvetica
+
            'axes.prop_cycle': default_color_cycler,
            'axes.labelsize': 8,
            'axes.linewidth': 1,
@@ -75,6 +72,15 @@ _params = {'font.family': 'sans-serif',
            'lines.markersize': 4,
            # 'lines.markeredgewidth': 0,
            # 0 will make line-type markers, such as '+', 'x', invisible
+
+           # Revert some properties to mpl v1 which is more suitable for publishing
+           'axes.autolimit_mode': 'round_numbers',
+           'axes.xmargin': 0,
+           'axes.ymargin': 0,
+           'xtick.direction': 'in',
+           'xtick.top': True,
+           'ytick.direction' : 'in',
+           'ytick.right': True,
            }
 
 acs_decorator = MPLdecorator(_params)