+++ mpltex/aps.py
@@ -4,7 +4,7 @@ aps.py
 ======
 
 Context decorator for producing figures which is ready to publish
-in American Chemical Society.
+in American Physical Society.
 
 EPS format for image file are used, because it is high quality and working
 in actual physical size rahter than pixel unit.
@@ -14,6 +14,7 @@ in actual physical size rahter than pixel unit.
 from .general import MPLdecorator
 from .colors import default_color_cycler
 from .layout import GOLDEN_RATIO
+from .styles import latex_preamble
 
 __all__ = ['aps_decorator', ]
 
@@ -34,11 +35,7 @@ _params = {'font.family': 'sans-serif',
            'font.size': 8,
            'text.usetex': True,
            # To force LaTeX use Helvetica fonts.
-           'text.latex.preamble': [r'\usepackage{siunitx}',
-                                   r'\sisetup{detect-all}',
-                                   r'\usepackage{helvet}',
-                                   r'\usepackage[eulergreek,EULERGREEK]{sansmath}',
-                                   r'\sansmath'],
+           'text.latex.preamble': latex_preamble,
            'axes.prop_cycle': default_color_cycler,
            'axes.labelsize': 8,
            'axes.linewidth': 1,
@@ -75,6 +72,15 @@ _params = {'font.family': 'sans-serif',
            'lines.markersize': 3,
            # 'lines.markeredgewidth' : 0,
            # 0 will make line-type markers, such as '+', 'x', invisible
+
+           # Revert some properties to mpl v1 which is more suitable for publishing
+           'axes.autolimit_mode': 'round_numbers',
+           'axes.xmargin': 0,
+           'axes.ymargin': 0,
+           'xtick.direction': 'in',
+           'xtick.top': True,
+           'ytick.direction' : 'in',
+           'ytick.right': True,
            }
 
 aps_decorator = MPLdecorator(_params)