+++ mpltex/web.py
@@ -22,6 +22,7 @@ screen, so that matplotlib will draw things nicely.
 from .general import MPLdecorator
 from .layout import GOLDEN_RATIO, point2inch
 from .colors import default_color_cycler
+from .styles import latex_preamble
 
 __all__ = ['web_decorator', ]
 
@@ -51,11 +52,7 @@ _params = {'font.family': 'sans-serif',
            'font.weight': 'normal',
            'text.usetex': True,
            # To force LaTeX use Helvetica fonts.
-           'text.latex.preamble': [r'\usepackage{siunitx}',
-                                   r'\sisetup{detect-all}',
-                                   r'\usepackage{helvet}',
-                                   r'\usepackage[eulergreek,EULERGREEK]{sansmath}',
-                                   r'\sansmath'],
+           'text.latex.preamble': latex_preamble,
            'axes.prop_cycle': default_color_cycler,
            'axes.labelsize': 'medium',
            'axes.labelweight': 'normal',
@@ -88,7 +85,6 @@ _params = {'font.family': 'sans-serif',
            'legend.columnspacing': 1,
 
            # 'text.fontsize' : 'medium',
-
            'xtick.major.size': 3,
            # 'xtick.minor.size': 2,
            'xtick.major.width': _line_width,
@@ -97,8 +93,6 @@ _params = {'font.family': 'sans-serif',
            # 'xtick.minor.pad': 4,
            # 'xtick.color' : k,
            'xtick.labelsize': 'medium',
-           # 'xtick.direction': 'in',
-
            'ytick.major.size': 3,
            # 'ytick.minor.size': 2,
            'ytick.major.width': _line_width,
@@ -107,12 +101,19 @@ _params = {'font.family': 'sans-serif',
            # 'ytick.minor.pad': 4,
            # 'ytick.color': k,
            'ytick.labelsize': 'medium',
-           # 'ytick.direction' : 'in',
-
            'lines.linewidth': _line_width,
            'lines.markersize': 3,
            # 'lines.markeredgewidth' : 0,
            # 0 will make line-type markers, such as '+', 'x', invisible
+
+           # Revert some properties to mpl v1 which is more suitable for publishing
+           'axes.autolimit_mode': 'round_numbers',
+           'axes.xmargin': 0,
+           'axes.ymargin': 0,
+           'xtick.direction': 'in',
+           'xtick.top': True,
+           'ytick.direction' : 'in',
+           'ytick.right': True,
            }
 
 web_decorator = MPLdecorator(_params)